refactor(dashboard): split agents.ts (773 LOC) into execution/ modules
Split the monolithic execution surface into focused modules:
- execution/shared.ts: signals, label functions, HandoffButtons, MonitorStat
- execution/queue.ts: ExecutionQueueBody, QueueCard
- execution/sessions.ts: SessionBriefsBody, SessionCard
- execution/operations.ts: OperationBriefsBody, OperationCard
- execution/workers.ts: WorkerSupportRow, ContinuityRow, LodgeTickCard, LodgeCheckinRow
- agents.ts: Execution shell (224 LOC, down from 773)

Also adds an "all clear" empty state when queue, sessions, and
operations are all empty, replacing individual "없습니다" messages
with a single green-tinted card.
